core: refactor: remove 'char' type
replace it to the string type holding single character.

+++ mod/test/case/core/syntax/FBOExprTest.cc
@@ -262,9 +262,9 @@ TEST_F(FBOExprTest, testLogicalBinaryOpChar) {
     make()
         .parse(R"SRC(
         main() int
-            a := 'l'
-            b := 'w'
-            a > b
+            a := "l"
+            b := "w"
+            (a as byte) > (b as byte)
     )SRC")
         .shouldVerified(true);
